requirements.txt
anticaptchaofficial==1.0.47
beautifulsoup4==4.11.1
jinja2==3.1.2
pre-commit==3.0.1
psycopg2==2.9.5
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==2.0.0
